requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
fastapi-mail
psycopg2-binary
